scene: preserve wide model indices and small transforms
Restore bounded authored UInt32 geometry admission through the existing prepared mesh and Metal index type. Preserve compact UInt16 upload and reject unknown, malformed, and out-of-range indices. Normalize inverse-transpose in Double to avoid rejecting small nonsingular authored transforms.

Validation: 14 reader/pipeline tests, 9 related model tests, Debug build and code health pass. Isolated 3662790108 admits layer 3694; final screenshot retains orbit/clock output. Whole scene remains NON-PASS and per-model completion plus sphere visual acceptance are still open.

+++ MyWallpaperX/Core/SteamWorkshopScene/Format/SceneMdlStaticModel.swift
@@ -25,7 +25,7 @@ nonisolated struct SceneMdlStaticModel: Equatable {
     let materialPath: String
     let bounds: Bounds
     let vertices: [Vertex]
-    let indices: [UInt16]
+    let indices: [UInt32]
 
     nonisolated var triangleCount: Int { indices.count / 3 }
 }
